Customize game and let others queue up to custom games, and other enhancements in tournaments

+++ backend/match/src/routes/http/game.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance } from "fastify";
 import credentialAuthCheck from "../../plugins/validateToken.js";
 import { GameSettings } from "../ws/session.js";
-import { checkMatch, createMatch, deleteMatch, getPlayerLevelAgainstAI, getUserMatches, notifyMatchParticipants } from "../../services/match.service.js";
+import { checkMatch, createMatch, deleteMatch, ensurePlayersInMatchesAreConnected, getPlayerLevelAgainstAI, getUserMatches, notifyMatchParticipants } from "../../services/match.service.js";
 // List of game modes
 // 1. 1v1
 // 2. 1vAI
@@ -11,6 +11,7 @@ import { checkMatch, createMatch, deleteMatch, getPlayerLevelAgainstAI, getUserM
 interface PreconfiguredGameSettings {
     mode: string;
     userIds?: string[];
+    matchId?: string;
 }
 
 export function gameHttpRoutes(app: FastifyInstance) {
@@ -34,11 +35,16 @@ export function gameHttpRoutes(app: FastifyInstance) {
         try {
             const body = request.body as PreconfiguredGameSettings;
             // get current pending match with same game type
+            const query: { gameType: string, status: string, id?: string } = {
+                gameType: body.mode,
+                status: 'pending',
+            }
+
+            if (body.matchId) {
+                query.id = body.matchId;
+            }
             const match = await app.prisma.match.findFirst({
-                where: {
-                    gameType: body.mode,
-                    status: 'pending',
-                },
+                where: query,
                 include: {
                     participants: true,
                 },
@@ -133,6 +139,9 @@ export function gameHttpRoutes(app: FastifyInstance) {
     app.post('/leave-queue', async (request, reply) => {
         const match = await checkMatch(request.user, app);
         if (match) {
+            if (match.tournamentId !== null && match.tournamentId !== '') {
+                return reply.status(400).send({ error: 'Cannot leave queue for tournament match' });
+            }
             await app.prisma.matchParticipant.deleteMany({ where: { matchId: match.id, userId: request.user } });
             // check if match is empty, delete it
             const participants = await app.prisma.matchParticipant.findMany({ where: { matchId: match.id } });
@@ -146,6 +155,8 @@ export function gameHttpRoutes(app: FastifyInstance) {
 
     // get is in queue
     app.get('/is-in-queue', async (request, reply) => {
+        await ensurePlayersInMatchesAreConnected(app, request.user);
+        console.log('is in queue');
         const match = await checkMatch(request.user, app);
         return reply.status(200).send({ message: 'In queue', inQueue: match !== null, since: match?.createdAt });
     });
@@ -180,5 +191,102 @@ export function gameHttpRoutes(app: FastifyInstance) {
             return reply.status(400).send({ error: 'Cannot delete match' });
         }
         return reply.status(200).send({ message: 'Match deleted', match });
+    });
+
+    app.get('/get-custom-games', async (request, reply) => {
+        try {
+            const matches = await app.prisma.match.findMany({
+                where: {
+                    gameType: 'custom',
+                    status: 'pending',
+                },
+                include: {
+                    participants: true,
+                },
+                orderBy: {
+                    createdAt: 'desc',
+                },
+            });
+            return reply.status(200).send({ matches });
+        } catch (error) {
+            console.error('Error getting custom games:', error);
+            return reply.status(500).send({ error: 'Internal server error' });
+        }
+    });
+
+
+
+    const isValidSettings = (settings: GameSettings) : string | null => {
+         // Validate settings
+         if (settings.players < 1 || settings.players > 4) {
+            return 'Invalid number of players';
+        }
+        if (settings.aiPlayers && (settings.aiPlayers < 0 || settings.aiPlayers > 3)) {
+            return 'Invalid number of AI players';
+        }
+        if (settings.timeLimit && settings.timeLimit < 60000) { // Minimum 1 minute
+            return 'Invalid time limit';
+        }
+        if (settings.balls && (settings.balls < 1 || settings.balls > 3)) {
+            return 'Invalid number of balls';
+        }
+        if (settings.aiLevel && (settings.aiLevel < 1 || settings.aiLevel > 9)) {
+            return 'Invalid AI level';
+        }
+        if (settings.winScore && settings.winScore < 1) {
+            return 'Invalid win score';
+        }
+        if (settings.timeLimit == 0 && settings.winScore == 0) {
+            return 'Cannot set both time limit and win score to 0';
+        }
+        if (settings.timeLimit && settings.timeLimit > 0 && settings.winScore && settings.winScore > 0) {
+            return 'Cannot set both time limit and win score';
+        }
+        if (settings.startScore && settings.startScore < 0) {
+            return 'Invalid start score';
+        }
+        if (settings.startScore && settings.winScore && settings.startScore > settings.winScore) {
+            return 'Start score cannot be greater than win score';
+        }
+        if (settings.startScore && (settings.startScore <= 0 && settings.terminatePlayers)) {
+            return 'Cannot set start score to 0 when terminating players';
+        }
+        return null;
+    }
+
+    app.post('/create-custom', async (request, reply) => {
+        try {
+            const settings = request.body as GameSettings;
+            const currentMatch = await checkMatch(request.user, app);
+            if (currentMatch) {
+                return reply.status(400).send({ error: 'You are already in a match / queue!' });
+            }
+            // Validate settings
+            const error = isValidSettings(settings);
+            if (error) {
+                return reply.status(400).send({ error });
+            }
+
+            // Create a new match with custom settings
+            const match = await app.prisma.match.create({
+                data: {
+                    userId: request.user,
+                    gameType: 'custom',
+                    status: 'pending',
+                    settings: settings,
+                    participants: {
+                        create: {
+                            userId: request.user,
+                            joinedAt: new Date(),
+                        },
+                    },
+                },
+            });
+            await notifyMatchParticipants(match.id, app);
+            return reply.status(200).send({ message: 'Custom game created successfully', match });
+        } catch (error) {
+            console.error('Error creating custom game:', error);
+            return reply.status(500).send({ error: 'Internal server error' });
+        }
     });
 }
